+++ src/main/java/com/maxmind/geoip2/exception/InvalidRequestException.java
@@ -1,8 +1,6 @@
 package com.maxmind.geoip2.exception;
 
-import java.net.MalformedURLException;
 import java.net.URI;
-import java.net.URL;
 
 /**
  * This class represents a non-specific error returned by MaxMind's GeoIP2 web
@@ -53,16 +51,4 @@ public final class InvalidRequestException extends GeoIp2Exception {
         return this.uri;
     }
 
-    /**
-     * @return the URL queried.
-     * @deprecated Use getUri() instead
-     */
-    @Deprecated
-    public URL getUrl() {
-        try {
-            return this.uri.toURL();
-        } catch (MalformedURLException e) {
-            return null;
-        }
-    }
 }